Require in-window Stop button to confirm MacWhisper recording
MacWhisper's status-menu extra can show stale "Stop Recording" and
"Recording <Platform> Meeting" items long after recording has actually
stopped — sometimes all three of "Record X Meeting", "Recording X
Meeting" and "Stop Recording" sit in the menu simultaneously. The old
performCheckRecording trusted the menu alone, so the upfront pre-check
in performStartRecording would short-circuit with "already recording"
when MacWhisper wasn't recording at all — the app showed "Recording"
but no file was captured.

The in-window Stop button (AXDescription "Stop", AXHelp "Stop the
meeting recording") only exists while MacWhisper is genuinely
recording, so it's the ground-truth signal. The fix:

- Tighten performCheckRecording to require BOTH menu signal AND
  in-window Stop button (split into menuExtraShowsRecording and
  anyWindowHasActiveStopButton helpers, moved to a +TruthCheck
  extension to stay under SwiftLint's body-length cap).
- Drop the upfront pre-check in performStartRecording. The menu-extra
  start path naturally handles the "MacWhisper is already recording"
  case (the "Record X Meeting" item is absent then), and the in-window
  fallback now checks the in-window Stop button before pressing the
  Record button, so we still don't trigger a duplicate when MacWhisper
  really is recording.
- Reuse windowHasActiveStopButton in performStopRecording's in-window
  stop scan.
